Example csv files that show the format of the input files.
There are two types of input files. Each Sakai institution typically would send in a pair of files, but sometimes they send in only one file.

sakai_events_example_input_file.csv is processed by the process_events_file.py python script. Good practice to rename the csv file to indicate which institution is providing, which also makes the file name unique. All the "events" files go into one folder. The python script parses all the files in the folder.

sakai_tools_placement_example_input_file.csv is processed by the process_tools_file.py python script. As in the other python script, rename the files to indicate the institution from which they come and place all the files from the different institutions into one folder.

You should have two folders, one for all the events csv files and a different folder for all the tools csv folders.

Also notice the last column in each file. This is a manually calculated column that needs to be there before running the script. It normalizes the data into the percent of each event or tool, therefore that last column contains all fractions which should add up to 100%. You need to add this column in yourself and re-save as csv, which is a lot of work and a pain.

The next iteration of the python scripts I intend to fix this issue, so that python does this calculation. Currently the python scripts only perform the aggregation of all the data, adding up the information in that last column across all the institutional data.