Updated the example files to remove a column that is no longer required.
In the previous version of these scripts, one had to manually update each file by calculating the total events (or tools placements, etc) and then the percentage of each event relative to the total. In the new version of the scripts that value is calculated by Python instead. No more manually updating.

As a result, the files are more in their original data format.